Remove cast from `libc::chmod()` call.
It's no longer necessary to case `self.mode` to `u16`, since it's now of
type `libc::mode_t`.

+++ src/changes.rs
@@ -63,7 +63,7 @@ impl<'o> ChMod<'o> {
             return Ok(());
         } else {
             unsafe {
-                match libc::chmod(self.file.c_path(), self.mode as u16) {
+                match libc::chmod(self.file.c_path(), self.mode) {
                     0 => Ok(()),
                     _ => return Err(Box::new(LibCError::from_errno())),
                 }
